+++ src/client/components/pages/Courses/MeetingModal.tsx
@@ -7,10 +7,6 @@ import {
   ModalFooter,
   ModalHeader,
   VARIANT,
-  BorderlessButton,
-  Dropdown,
-  TextInput,
-  ValidationErrorMessage,
 } from 'mark-one';
 import React, {
   FunctionComponent,
@@ -18,17 +14,10 @@ import React, {
   Ref,
   useEffect,
   useRef,
-  useState,
 } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faEdit, faTrash } from '@fortawesome/free-solid-svg-icons';
 import styled from 'styled-components';
-import { VerticalSpace } from 'client/components/layout';
-import { ButtonLayout, ListLayout, TimePicker } from 'client/components/general';
-import DAY, { dayEnumToString, days } from 'common/constants/day';
-import { calculateStartEndTimes, convert12To24HourTime } from 'common/utils/timeHelperFunctions';
-import { meetingTimeSlots } from 'common/constants/timeslots';
 import { instructorDisplayNameToFirstLast } from '../utils/instructorDisplayNameToFirstLast';
+import { MeetingTimesList } from './MeetingTimesList';
 
 /**
  * A component that applies styling for text that indicates the faculty has
@@ -116,14 +105,6 @@ const RoomAvailabilityBody = styled.div`
   overflow: auto;
 `;
 
-/**
- * Contains the meeting day and time selectors so they can be displayed side by side
- */
-const TimeSelector = styled.div`
-  display: flex;
-  flex-direction: row;
-`;
-
 const MeetingModal: FunctionComponent<MeetingModalProps> = function ({
   isVisible,
   onClose,
@@ -151,63 +132,6 @@ const MeetingModal: FunctionComponent<MeetingModalProps> = function ({
   const semKey = term.toLowerCase() as TermKey;
   const instance = course[semKey];
 
-  /**
-   * Keeps track of the current meetings for this instance. This is updated as
-   * users add and edit meetings in the modal.
-   */
-  const [
-    currentMeetings,
-    setCurrentMeetings,
-  ] = useState(instance.meetings);
-
-  /**
-   * The meeting within the list of meetings that is currently being edited
-   */
-  const [
-    currentMeetingId,
-    setCurrentMeetingId,
-  ] = useState(null as string);
-
-  /**
-   * The selected day in the dropdown for the meeting currently being edited
-   */
-  const [
-    currentDay,
-    setCurrentDay,
-  ] = useState(null as DAY);
-
-  /**
-   * The timeslot currently selected in the time picker dropdown
-   */
-  const [
-    currentTimeslot,
-    setCurrentTimeslot,
-  ] = useState(null as string);
-
-  /**
-   * The start time value for the meeting currently being edited
-   */
-  const [
-    currentStartTime,
-    setCurrentStartTime,
-  ] = useState(null as string);
-
-  /**
-   * The end time value for the meeting currently being edited
-   */
-  const [
-    currentEndTime,
-    setCurrentEndTime,
-  ] = useState(null as string);
-
-  /**
-   * The current value of the error message when creating or editing a meeting time
-   */
-  const [
-    meetingTimeError,
-    setMeetingTimeError,
-  ] = useState('');
-
   return (
     <Modal
       ariaLabelledBy="editMeeting"
@@ -225,185 +149,7 @@ const MeetingModal: FunctionComponent<MeetingModalProps> = function ({
           <MeetingScheduler>
             <MeetingSchedulerHeader>{`Meeting times for ${course.catalogNumber}`}</MeetingSchedulerHeader>
             <MeetingSchedulerBody>
-              <div className="meeting-times-section">
-                <ul>
-                  {currentMeetings.map(
-                    (meeting) => (
-                      <ListLayout key={meeting.id}>
-                        <BorderlessButton
-                          id={`deleteButton${meeting.id}`}
-                          variant={VARIANT.DANGER}
-                          onClick={
-                            (): void => {}
-                          }
-                        >
-                          <FontAwesomeIcon icon={faTrash} />
-                        </BorderlessButton>
-                        {
-                          meeting.id === currentMeetingId
-                            ? (
-                              <div>
-                                <TimeSelector>
-                                  <Dropdown
-                                    id="meetingDay"
-                                    name="meetingDay"
-                                    label="Meeting Day"
-                                    options={
-                                      days.map((day) => ({
-                                        value: day,
-                                        label: dayEnumToString(day),
-                                      }))
-                                    }
-                                    value={currentDay}
-                                    onChange={
-                                      (event
-                                      : React.ChangeEvent<HTMLSelectElement>)
-                                      : void => {
-                                        setCurrentDay(
-                                          event.currentTarget.value as DAY
-                                        );
-                                      }
-                                    }
-                                    hideError
-                                    isRequired
-                                    isLabelVisible={false}
-                                  />
-                                  <TimePicker
-                                    id="timeslots"
-                                    name="timeslots"
-                                    options={[{ value: '', label: '🕒' }]
-                                      .concat(meetingTimeSlots.map((slot) => ({
-                                        value: slot,
-                                        label: slot,
-                                      })))}
-                                    value={currentTimeslot}
-                                    onChange={
-                                      (event
-                                      : React.ChangeEvent<HTMLSelectElement>)
-                                      : void => {
-                                        if (event.currentTarget.value !== '') {
-                                          const times = calculateStartEndTimes(
-                                            event.currentTarget.value
-                                          );
-                                          setCurrentStartTime(times.start);
-                                          setCurrentEndTime(times.end);
-                                        }
-                                      }
-                                    }
-                                  />
-                                  <TextInput
-                                    id="meetingStartTime"
-                                    name="meetingStartTime"
-                                    label="Meeting Start Time"
-                                    type="time"
-                                    value={convert12To24HourTime(
-                                      currentStartTime
-                                    )}
-                                    onChange={
-                                      (event
-                                      : React.ChangeEvent<HTMLInputElement>)
-                                      : void => {
-                                        setCurrentStartTime(
-                                          event.currentTarget.value
-                                        );
-                                      }
-                                    }
-                                    hideError
-                                    isRequired
-                                    isLabelVisible={false}
-                                  />
-                                  <TextInput
-                                    id="meetingEndTime"
-                                    name="meetingEndTime"
-                                    label="Meeting End Time"
-                                    type="time"
-                                    value={convert12To24HourTime(
-                                      currentEndTime
-                                    )}
-                                    onChange={
-                                      (event
-                                      : React.ChangeEvent<HTMLInputElement>)
-                                      : void => {
-                                        setCurrentEndTime(
-                                          event.currentTarget.value
-                                        );
-                                      }
-                                    }
-                                    hideError
-                                    isRequired
-                                    isLabelVisible={false}
-                                  />
-                                  <ValidationErrorMessage
-                                    id="meetingTimeErrorMessage"
-                                  >
-                                    {meetingTimeError}
-                                  </ValidationErrorMessage>
-                                </TimeSelector>
-                                <div>
-                                  <span>
-                                    Room:
-                                    {` ${meeting.room.name}`}
-                                  </span>
-                                  <ButtonLayout>
-                                    <Button
-                                      id="closeButton"
-                                      onClick={
-                                        (): void => {}
-                                      }
-                                      variant={VARIANT.SECONDARY}
-                                    >
-                                      Close
-                                    </Button>
-                                    <Button
-                                      id="showRoomsButton"
-                                      onClick={
-                                        (): void => {}
-                                      }
-                                      variant={VARIANT.PRIMARY}
-                                    >
-                                      Show Rooms
-                                    </Button>
-                                  </ButtonLayout>
-                                </div>
-                              </div>
-                            )
-                            : (
-                              <>
-                                <span>{`${dayEnumToString(meeting.day)}, ${meeting.startTime} to ${meeting.endTime} in ${meeting.room.name}`}</span>
-                                <BorderlessButton
-                                  id={`editButton${meeting.id}`}
-                                  variant={VARIANT.INFO}
-                                  onClick={
-                                    (): void => {
-                                      setCurrentMeetingId(meeting.id);
-                                      setCurrentDay(meeting.day);
-                                      setCurrentStartTime(meeting.startTime);
-                                      setCurrentEndTime(meeting.endTime);
-                                    }
-                                  }
-                                >
-                                  <FontAwesomeIcon icon={faEdit} />
-                                </BorderlessButton>
-                              </>
-                            )
-                        }
-                      </ListLayout>
-                    )
-                  )}
-                </ul>
-                <VerticalSpace />
-                <ButtonLayout>
-                  <Button
-                    id="addNewTimeButton"
-                    onClick={
-                      (): void => {}
-                    }
-                    variant={VARIANT.SECONDARY}
-                  >
-                    Add New Time
-                  </Button>
-                </ButtonLayout>
-              </div>
+              <MeetingTimesList meetings={instance.meetings} />
               <h3>
                 Faculty Notes
               </h3>
